+++ src/TypeUtils.jl
@@ -41,7 +41,7 @@ if !isdefined(Base, :get_extension)
 end
 
 """
-    @public args...
+    TypeUtils.@public args...
 
 declares `args...` as being `public` even though they are not exported. For Julia version
 < 1.11, this macro does nothing. Using this macro also avoid errors with CI and coverage
@@ -51,6 +51,7 @@ tools.
 macro public(args::Union{Symbol,Expr}...)
     VERSION ≥ v"1.11.0-DEV.469" ? esc(Expr(:public, args...)) : nothing
 end
+VERSION ≥ v"1.11.0-DEV.469" && eval(Expr(:public, Symbol("@public")))
 
 """
     TypeUtils.Dim