+++ NEWS.md
@@ -1,5 +1,7 @@
 # User visible changes in `TypeUtils`
 
+- Non-exported `TypeUtils.@public` macro is now public.
+
 # Version 1.4.1
 
 - `AbstractUnitRange{<:Integer}` has been replaced by `AbstractRange{<:Integer}` as the
@@ -13,8 +15,8 @@
   `ArrayShape{N}` is restricted to integers and integer-valued unit-ranges which better
   correspond to Julia's way of representing an `N`-dimensional array shape.
 
-- New macro `@public` used to declare non-exported public symbols. Does nothing for Julia
-  versions older than 1.11.
+- New macro `TypeUtils.@public` used to declare non-exported public symbols. Does nothing
+  for Julia versions older than 1.11.
 
 - For other packages, it may be sufficient to extend `convert_eltype(T, X)` to a given
   type `X` to have `convert_eltype(T, x::X)` work for instances `x` of that type.