+++ NEWS.md
@@ -1,5 +1,7 @@
 # User visible changes in `TypeUtils`
 
+# Version 1.4.2
+
 - New `nearest(T,x)` method to return the value of type `T` that is the nearest to `x`.
   For `T` integer and `x` real, it can be seen as rounding with clamping to avoid
   overflows.